requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.0
jinja2==3.1.2
python-multipart==0.0.6
email-validator==2.1.0